feat: search is moved to library
