+++ tests/negascout_test.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include "src/lib/board.hpp"
 #include "utility.hpp"
-#include "search.h"
+#include "src/lib/search.h"
 
 
 class NegascoutTest : public ::testing::Test {
